code/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements/requirements.in
#
anyio==3.5.0
    # via jupyter-server
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asttokens==2.0.5
    # via stack-data
attrs==21.4.0
    # via jsonschema
babel==2.10.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.11.1
    # via nbconvert
bleach==5.0.0
    # via nbconvert
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
click==8.1.3
    # via
    #   nltk
    #   psaw
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
entrypoints==0.4
    # via
    #   jupyter-client
    #   nbconvert
executing==0.8.3
    # via stack-data
fastjsonschema==2.15.3
    # via nbformat
idna==3.3
    # via
    #   anyio
    #   requests
importlib-metadata==4.11.3
    # via jupyterlab-server
ipykernel==6.13.0
    # via notebook
ipython==8.3.0
    # via
    #   ipykernel
    #   jupyterlab
ipython-genutils==0.2.0
    # via notebook
jedi==0.18.1
    # via ipython
jinja2==3.1.2
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via nltk
json5==0.9.6
    # via jupyterlab-server
jsonschema==4.4.0
    # via
    #   jupyterlab-server
    #   nbformat
jupyter-client==7.3.0
    # via
    #   ipykernel
    #   jupyter-server
    #   nbclient
    #   notebook
jupyter-core==4.10.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbconvert
    #   nbformat
    #   notebook
jupyter-server==1.17.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   notebook-shim
jupyterlab==3.3.4
    # via -r requirements/requirements.in
jupyterlab-pygments==0.2.2
    # via nbconvert
jupyterlab-server==2.13.0
    # via jupyterlab
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
nbclassic==0.3.7
    # via jupyterlab
nbclient==0.6.0
    # via nbconvert
nbconvert==6.5.0
    # via
    #   jupyter-server
    #   notebook
nbformat==5.3.0
    # via
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.5
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
nltk==3.7
    # via -r requirements/requirements.in
notebook==6.4.11
    # via nbclassic
notebook-shim==0.1.0
    # via nbclassic
numpy==1.22.3
    # via pandas
packaging==21.3
    # via
    #   ipykernel
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
pandas==1.4.2
    # via -r requirements/requirements.in
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
praw==7.5.0
    # via -r requirements/requirements.in
prawcore==2.3.0
    # via praw
prometheus-client==0.14.1
    # via
    #   jupyter-server
    #   notebook
prompt-toolkit==3.0.29
    # via ipython
psaw==0.1.0
    # via -r requirements/requirements.in
psutil==5.9.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pure-eval==0.2.2
    # via stack-data
pycparser==2.21
    # via cffi
pygments==2.12.0
    # via
    #   ipython
    #   nbconvert
pyparsing==3.0.8
    # via packaging
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
python-dotenv==0.20.0
    # via -r requirements/requirements.in
pytz==2022.1
    # via
    #   babel
    #   pandas
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   notebook
regex==2022.4.24
    # via nltk
requests==2.27.1
    # via
    #   -r requirements/requirements.in
    #   jupyterlab-server
    #   prawcore
    #   psaw
    #   update-checker
send2trash==1.8.0
    # via
    #   jupyter-server
    #   notebook
six==1.16.0
    # via
    #   asttokens
    #   bleach
    #   python-dateutil
sniffio==1.2.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
stack-data==0.2.0
    # via ipython
terminado==0.13.3
    # via
    #   jupyter-server
    #   notebook
tinycss2==1.1.1
    # via nbconvert
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
tqdm==4.64.0
    # via nltk
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
update-checker==0.18.0
    # via praw
urllib3==1.26.9
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via
    #   bleach
    #   tinycss2
websocket-client==1.3.2
    # via
    #   jupyter-server
    #   praw
zipp==3.8.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
